Util update: Merge SimpleDownloader into ContentDownloader
Allows for ContentDownloader to be the one util to be used to
download from arbitrary URLs.
