+++ lib/java-extras/src/main/java/org/triplea/io/ContentDownloader.java
@@ -6,14 +6,20 @@ import java.io.InputStream;
 import java.net.URI;
 import java.util.Optional;
 import java.util.function.Consumer;
+import java.util.function.Function;
+import javax.annotation.Nullable;
 import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.http.HttpEntity;
+import org.apache.http.HttpHost;
 import org.apache.http.HttpStatus;
+import org.apache.http.client.config.RequestConfig;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.triplea.java.Interruptibles;
+import org.triplea.java.StringUtils;
 
 /**
  * Downloads content from HTTP resources. Meant to be used with a try-with-resources block, eg:
@@ -32,6 +38,7 @@ import org.triplea.java.Interruptibles;
  * <p>Warning: The input stream provided by this class can only be consumed once (property of input
  * streams, the input stream is not reset in any way after reading it).
  */
+@Slf4j
 public final class ContentDownloader implements CloseableDownloader {
   private final CloseableHttpClient httpClient;
 
@@ -93,5 +100,94 @@ public final class ContentDownloader implements CloseableDownloader {
     stream.close();
     response.close();
     httpClient.close();
+  }
+
+  public static Optional<String> downloadAsString(final URI uri) {
+    return downloadAsStringWithProxy(uri, null);
+  }
+
+  public static Optional<String> downloadAsStringWithProxy(
+      final URI uri, @Nullable final HttpHost proxy) {
+    try {
+      return Optional.of(download(uri, StringUtils::readFully, proxy));
+    } catch (final DownloadException e) {
+      if (e.isServerError()) {
+        log.error(
+            "Contact TripleA for support, Error downloading: {}, "
+                + "server error. Server status: {}, error: {}",
+            uri,
+            e.getStatusCode(),
+            e.getMessage());
+        return Optional.empty();
+      } else {
+        log.warn(
+            "Check internent connection. Error downloading: {}, status: {}, error: {}",
+            uri,
+            e.getStatusCode(),
+            e.getMessage());
+        return Optional.empty();
+      }
+    }
+  }
+
+  /**
+   * Downloads content from a given URI and runs a function on the downloaded content and returns
+   * that functions output.
+   *
+   * @param uri The URI to download
+   * @param streamProcessor Function to process the downloaded input.
+   * @return Result of the stream processing function
+   * @throws DownloadException Thrown if there are any problems during download.
+   */
+  private static <T> T download(
+      final URI uri, final Function<InputStream, T> streamProcessor, @Nullable final HttpHost proxy)
+      throws DownloadException {
+
+    final HttpGet request = new HttpGet(uri);
+    Optional.ofNullable(proxy)
+        .ifPresent(
+            proxyHost ->
+                request.setConfig(
+                    RequestConfig.copy(request.getConfig()).setProxy(proxyHost).build()));
+
+    try (CloseableHttpClient httpClient = HttpClients.custom().disableCookieManagement().build();
+        CloseableHttpResponse response = httpClient.execute(request)) {
+      final int statusCode = response.getStatusLine().getStatusCode();
+      if (statusCode != HttpStatus.SC_OK) {
+        throw new DownloadException(
+            statusCode, String.format("Unexpected status code (%d)", statusCode));
+      }
+      final HttpEntity entity =
+          Optional.ofNullable(response.getEntity())
+              .orElseThrow(() -> new DownloadException(statusCode, "Entity is missing"));
+      try (InputStream stream = entity.getContent()) {
+        return streamProcessor.apply(stream);
+      } catch (final IOException e) {
+        throw new DownloadException(statusCode, e);
+      }
+    } catch (final IOException e) {
+      throw new DownloadException(0, e);
+    }
+  }
+
+  /** Thrown if there are any problems downloading or reading downloaded content. */
+  @Getter
+  public static class DownloadException extends Exception {
+    private static final long serialVersionUID = -4049197878908223175L;
+    private final int statusCode;
+
+    DownloadException(final int statusCode, final String message) {
+      super(message);
+      this.statusCode = statusCode;
+    }
+
+    DownloadException(final int statusCode, final Exception cause) {
+      super(cause);
+      this.statusCode = statusCode;
+    }
+
+    boolean isServerError() {
+      return statusCode >= 500;
+    }
   }
 }
